Biblioteca_Backend22-main/pyproject.toml
[project]
name = "biblioteca"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "luansilva25", email = "luansilva250807@gmail.com"},
]
dependencies = ["django>=5.1.6", "djangorestframework>=3.15.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
